+++ slavv_python/pipeline/edges/radius_utils.py
@@ -7,7 +7,7 @@ import numpy as np
 
 def _scalar_radius(radius_value: np.ndarray | float | int) -> float:
     """Convert isotropic or axis-aware radii into a single tracing radius."""
-    radius_array = np.asarray(radius_value, dtype=np.float32).reshape(-1)
+    radius_array = np.asarray(radius_value, dtype=np.float64).reshape(-1)
     if radius_array.size == 0:
         return 0.0
     if radius_array.size == 1:
